+++ Library/Source/Nano/System/NDataCipher.cpp
@@ -324,7 +324,7 @@ NStatus NDataCipher::DES3_Decrypt(NData &theData)
 //----------------------------------------------------------------------------
 NStatus NDataCipher::Blowfish_Encrypt(NData &theData)
 {	NIndex					n, numWords;
-	UInt32					*dataPtr;
+	unsigned long			*dataPtr;
 	blowfish_context_t		bfState;
 
 
@@ -336,8 +336,8 @@ NStatus NDataCipher::Blowfish_Encrypt(NData &theData)
 
 
 	// Get the state we need
-	dataPtr  = (UInt32 *) theData.GetData();
-	numWords =            theData.GetSize() / sizeof(UInt32);
+	dataPtr  = (unsigned long *) theData.GetData();
+	numWords =                   theData.GetSize() / sizeof(unsigned long);
 
 
 
@@ -361,7 +361,7 @@ NStatus NDataCipher::Blowfish_Encrypt(NData &theData)
 //----------------------------------------------------------------------------
 NStatus NDataCipher::Blowfish_Decrypt(NData &theData)
 {	NIndex					n, numWords;
-	UInt32					*dataPtr;
+	unsigned long			*dataPtr;
 	blowfish_context_t		bfState;
 
 
@@ -373,8 +373,8 @@ NStatus NDataCipher::Blowfish_Decrypt(NData &theData)
 
 
 	// Get the state we need
-	dataPtr  = (UInt32 *) theData.GetData();
-	numWords =            theData.GetSize() / sizeof(UInt32);
+	dataPtr  = (unsigned long *) theData.GetData();
+	numWords =                   theData.GetSize() / sizeof(unsigned long);
